requirements.txt
langgraph
langchain-openai
langchain-core
pdfplumber
python-dotenv
pydantic
openai
streamlit
